feat(energy): add 1-minute time range support and a range selection dropdown on the dashboard

+++ apps/api/backend/energy/schemas.py
@@ -18,7 +18,7 @@ class EnergyPoint(BaseModel):
 
 
 class EnergySummary(BaseModel):
-    range_hours: int
+    range_hours: float
     total_power_w: float
     total_kwh: float
     devices: list[EnergyDevice]
